Solve input quantity issues

+++ Controllers/AddIrrigation.swift
@@ -45,7 +45,7 @@ extension AddInterventionViewController: UITextFieldDelegate, CustomPickerViewPr
 
     self.irrigationUnitButton.setTitle(selectedValue, for: .normal)
     irrigationLabel.text = String(format: "Volume • %g %@", volume, unit)
-    updateInfoLabel(Double(volume), unit)
+    updateInfoLabel(volume, unit)
     irrigationPickerView.isHidden = true
     dimView.isHidden = true
   }
@@ -68,10 +68,10 @@ extension AddInterventionViewController: UITextFieldDelegate, CustomPickerViewPr
     let unit = irrigationUnitButton.titleLabel!.text!
 
     irrigationLabel.text = String(format: "Volume • %g %@", volume, unit)
-    updateInfoLabel(Double(volume), unit)
+    updateInfoLabel(volume, unit)
   }
 
-  private func updateInfoLabel(_ volume: Double, _ unit: String) {
+  private func updateInfoLabel(_ volume: Float, _ unit: String) {
     if volume == 0 {
       irrigationInfoLabel.text = "Le volume ne peut être nul"
       irrigationInfoLabel.textColor = AppColor.TextColors.Red
@@ -79,7 +79,7 @@ extension AddInterventionViewController: UITextFieldDelegate, CustomPickerViewPr
       irrigationInfoLabel.text = "Aucune culture sélectionnée"
       irrigationInfoLabel.textColor = AppColor.TextColors.Red
     } else {
-      let efficiency = Double(volume) / cropsView.totalSurfaceArea
+      let efficiency = volume / cropsView.totalSurfaceArea
 
       irrigationInfoLabel.text = String(format: "Soit %.1f %@ par hectare", efficiency, unit)
       irrigationInfoLabel.textColor = AppColor.TextColors.DarkGray
